feat: implement texture mapping system and stb_image integration
- Added Texture and TextureManager classes with static caching.
- Integrated stb_image as a static library (.a) via CMake sub-project.
- Refactored asset structure: moved shaders and textures to root /assets.
- Updated VertexArray and layout to support UV (texture) coordinates.
- Enhanced Fragment Shader to support texture sampling with color tinting.
- Fixed 64-bit pointer casting warnings in VertexArray.
- Added Timer utility for performance monitoring and animation logic.

+++ src/core/VertexArray.cpp
@@ -27,7 +27,7 @@ void VertexArray::addBuffer(const VertexBuffer& vb, const VertexBufferLayout& la
             element.type,
             element.normalize,
             layout.getStride(),
-            (const void*) offset
+            (const void*) (size_t)offset
         ));
 
         offset += element.count * VertexBufferElement::getSizeOfType(element.type);
